Make gtk-redshift read output of the redshift child process.
This allows gtk-redshift to show the current state of the redshift
process. gtk-redshift follows the enable state of redshift and toggles
the icon accordingly. Temperature etc. can also be followed but is not
implemented yet.

The implementation is changed to use glib to spawn the child process
instead of relying on python subprocess module. This is necessary
because of inflexibility in the python module.